backend/services/admin/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyJWT>=2.8.0
# bcrypt via packages/common (admin auth uses packages.common.src.auth, not passlib)
redis>=5.0.0
python-multipart>=0.0.6
sentry-sdk[fastapi]>=2.0.0
slowapi>=0.1.9
prometheus-client>=0.20.0
